+++ Gulpfile.js
@@ -7,10 +7,8 @@
  *
 **/
 
-
-// Require.
 var gulp = require('gulp');
-
+var package = require('./package.json');
 
 /**
  *
@@ -23,12 +21,10 @@ var gulp = require('gulp');
  *
 **/
 gulp.task('css', function(){
-	// Require.
-	var sass = require('gulp-sass');
-	// Require.
+	var sass         = require('gulp-sass');
 	var autoprefixer = require('gulp-autoprefixer');
 
-	return gulp.src('./src/dist/css/scss/build.scss')
+	return gulp.src(package.config.css.src)
         .pipe(sass({
 	        outputStyle: 'expanded',
         	errLogToConsole: true
@@ -37,5 +33,5 @@ gulp.task('css', function(){
             browsers: ['last 2 version', 'safari 5', 'ie 8', 'ie 9', 'opera 12.1', 'ios 6', 'android 4'],
             cascade: false
         }))
-		.pipe(gulp.dest('./src/dist/css'));
+		.pipe(gulp.dest(package.config.css.destDir));
 });